- Added an option to override the session cookie domain.  - fix: Search results not showing synopses  - fix: Redirects get displayed inside page layouts as '0'  - fix: Mysterious "0" Appearing When Admin Is Off  - fix: Deletion of Products

+++ lib/WebGUI/Operation/ProductManager.pm
@@ -673,7 +673,7 @@ sub www_listProducts {
 	while ($row = $sth->hashRef) {
 		$output .= '<tr>';
 		$output .= '<td>';
-		$output .= $session->icon->delete('op=deleteProduct;productId='.$row->{productId});
+		$output .= $session->icon->delete('op=deleteProduct;productId='.$row->{productId}, undef, $i18n->get("confirm delete product"));
 		$output .= $session->icon->edit('op=manageProduct;productId='.$row->{productId});
 		$output .= '</td>';
 		$output .= '<td>'.$row->{title}.'</td>';
